@@ -12,13 +12,13 @@
   "dependencies": {
     "@babel/runtime": "^7.5.5",
     "body-parser": "^1.19.0",
-    "core-js": "^3.1.4",
+    "core-js": "^3.2.1",
     "dayjs": "^1.8.15",
     "express": "^4.17.1",
     "path-to-regexp": "^3.0.0",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "react-hot-loader": "^4.12.10",
+    "react": "^16.9.0",
+    "react-dom": "^16.9.0",
+    "react-hot-loader": "^4.12.11",
     "react-redux": "^7.1.0",
     "react-router": "^5.0.1",
     "react-router-dom": "^5.0.1",
@@ -41,24 +41,24 @@
     "@babel/preset-env": "^7.5.5",
     "@babel/preset-react": "^7.0.0",
     "@babel/preset-typescript": "^7.3.3",
-    "@mdi/svg": "^3.9.97",
-    "@storybook/addon-actions": "^5.1.10",
-    "@storybook/addon-links": "^5.1.10",
-    "@storybook/addons": "^5.1.10",
-    "@storybook/react": "^5.1.10",
+    "@mdi/svg": "^4.0.96",
+    "@storybook/addon-actions": "^5.1.11",
+    "@storybook/addon-links": "^5.1.11",
+    "@storybook/addons": "^5.1.11",
+    "@storybook/react": "^5.1.11",
     "@testing-library/jest-dom": "^4.0.0",
-    "@testing-library/react": "^8.0.7",
+    "@testing-library/react": "^9.1.1",
     "@types/express": "^4.17.0",
     "@types/jest": "^24.0.17",
-    "@types/react": "^16.8.24",
+    "@types/react": "^16.9.1",
     "@types/react-dom": "^16.8.5",
     "@types/react-redux": "^7.1.1",
     "@types/react-router": "^5.0.3",
     "@types/react-router-dom": "^4.3.4",
     "@types/storybook__addon-actions": "^3.4.3",
     "@types/storybook__react": "^4.0.2",
-    "@typescript-eslint/eslint-plugin": "^1.13.0",
-    "@typescript-eslint/parser": "^1.13.0",
+    "@typescript-eslint/eslint-plugin": "^2.0.0",
+    "@typescript-eslint/parser": "^2.0.0",
     "autoprefixer": "^9.6.1",
     "babel-eslint": "^10.0.2",
     "babel-jest": "^24.8.0",
@@ -67,17 +67,15 @@
     "copy-webpack-plugin": "^5.0.4",
     "cross-env": "^5.2.0",
     "css-loader": "^3.2.0",
-    "cucumber": "^5.1.0",
-    "cucumber-messages": "^3.0.2",
     "eslint": "^6.1.0",
-    "eslint-config-airbnb-base": "^13.2.0",
+    "eslint-config-airbnb-base": "^14.0.0",
     "eslint-loader": "^2.2.1",
     "eslint-plugin-import": "^2.18.2",
-    "eslint-plugin-jest": "^22.14.1",
+    "eslint-plugin-jest": "^22.15.1",
     "eslint-plugin-react": "^7.14.3",
-    "eslint-plugin-react-hooks": "^1.6.1",
-    "file-loader": "^4.1.0",
-    "gherkin-testcafe": "^2.3.3",
+    "eslint-plugin-react-hooks": "^1.7.0",
+    "eslint-plugin-testcafe": "^0.2.1",
+    "file-loader": "^4.2.0",
     "html-webpack-plugin": "^3.2.0",
     "identity-obj-proxy": "^3.0.0",
     "jest": "^24.8.0",
@@ -89,7 +87,7 @@
     "postcss-loader": "^3.0.0",
     "prettier-stylelint": "^0.4.2",
     "redux-logger": "^3.0.6",
-    "sass-loader": "^7.1.0",
+    "sass-loader": "^7.2.0",
     "style-loader": "^1.0.0",
     "stylelint": "^10.1.0",
     "stylelint-config-standard": "^18.3.0",
@@ -97,12 +95,12 @@
     "stylelint-scss": "^3.9.3",
     "stylelint-selector-bem-pattern": "^2.1.0",
     "stylelint-webpack-plugin": "^0.10.5",
-    "testcafe": "^1.3.3",
+    "testcafe": "^1.4.1",
     "typescript": "^3.5.3",
-    "webpack": "^4.39.1",
+    "webpack": "^4.39.2",
     "webpack-bundle-analyzer": "^3.4.1",
     "webpack-cli": "^3.3.6",
-    "webpack-dev-server": "^3.7.2",
+    "webpack-dev-server": "^3.8.0",
     "webpack-manifest-plugin": "^2.0.4",
     "webpack-node-externals": "^1.7.2"
   },
@@ -124,8 +122,5 @@
     "lint-fix": "eslint -c ./config/linters/.eslintrc.json --ext tsx,ts src --fix",
     "build-storybook": "build-storybook -o public",
     "e2e": "node testcafe"
-  },
-  "resolutions": {
-    "gherkin": "joscha/gherkin-javascript#joscha/fix-exefile-constructor"
   }
 }
